Move dev-session files into dev-session folder
