+++ packages/app/src/cli/services/dev/processes/dev-session/dev-session.ts
@@ -1,7 +1,7 @@
 import {DevSessionLogger} from './dev-session-logger.js'
+import {DevSessionStatusManager} from './dev-session-status-manager.js'
+import {DevSessionProcessOptions} from './dev-session-process.js'
 import {AppEvent, AppEventWatcher} from '../../app-events/app-event-watcher.js'
-import {DevSessionStatusManager} from '../dev-session-status-manager.js'
-import {DevSessionProcessOptions} from '../dev-session.js'
 import {getExtensionUploadURL} from '../../../deploy/upload.js'
 import {outputDebug} from '@shopify/cli-kit/node/output'
 import {endHRTimeInMs, startHRTime} from '@shopify/cli-kit/node/hrtime'
